Add supports for DWIN_SET v0.4.5 + bug fix + add comments
Adds background support for the new Firmware Retraction pages, including the ability to return to the calling page, since it can be accessed from at least two places.
Added more comments to the code, to make it easier to understand and modify.
Limited files scrolling to a minimum index of -5, to solve a crashing bug on the Print screen.
Updated the release notes.