Renamed the bed_mesh macros and baselined distribution version of bed_mesh profiles in printer.cfg
Decided to rename the bed_mesh profiles to be filament-neutral and more informative.
Corrected the macros, accordingly.
Also decided upon a single non-zero cell in each profile, as distributed.  Klipper would not work correctly if the meshes were ALL zero, so one cell is non-zero in the smallest digit.  Results in (I hope) a visually obvious clue that the bed mesh profile needs to be calibrated on the host printer, as part of the installation and setup.