Found and fixed nasty bug in 0.4.7
Klipper deletes items from variables.cfg when we reboot.  When the unload_filament_after_print flag was deleted, it caused the printer to shut down any time that DGUS-Reloaded tried to display that button.
Refactored the function to use presets.cfg instead of variables.cfg to make that flag persistent.