Renamed preset storage and retrieval scripts for clarity. Changed default of Auto Unload to "disabled"
Since some of the presets.presets are now not material presets, I renamed the two functions in t5uid1.py, and everywhere they are called in the .cfg files, to improve readability of the code.

I also restored the default behaviour of Auto Unload After Print to Disabled, for consistency with v0.4.7.