Tweaked the _load_macro_menus method to test the cfg file's timestamp before loading the dictionary
This allows a user to edit and save the DGUS_Macro_Menus.cfg file without restarting the firmware.
The firmware loads the dictionary once at boot time, and then loads the macro menu contents from the dictionary.  If the dictionary is subsequently modified, the firmware will reload the dictionary, next time a macros page is selected.